+++ typings/classes/Utils.d.ts
@@ -1,9 +1,0 @@
-declare class Utils {
-    /**
-    * Checks for is the item object and returns it.
-    * @param {any} item The item to check.
-    * @returns {boolean} Is the item object or not.
-    */
-    isObject<K>(item: K): boolean;
-}
-export = Utils;