+++ typings/interfaces/MongoDatabaseEvents.d.ts
@@ -1,7 +1,0 @@
-import { Collection, Document, MongoClient } from 'mongodb';
-interface DatabaseEvents {
-    connecting: void;
-    ready: Collection<Document>;
-    destroy: MongoClient;
-}
-export = DatabaseEvents;