+++ typings/src/classes/DatabaseError.d.ts
@@ -1,11 +1,10 @@
 /**
  * DatabaseError class.
  */
-declare class DatabaseError extends Error {
+export declare class DatabaseError extends Error {
     /**
      * Creates a 'DatabaseError' instance.
      * @param {string} msg Error message.
      */
     constructor(msg?: string);
 }
-export = DatabaseError;